+++ packages/app-cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @happyvertical/smrt-app-cli
 
+## 1.0.0
+
+### Patch Changes
+
+- @happyvertical/smrt-users@1.0.0
+
+## 0.27.42
+
+### Patch Changes
+
+- @happyvertical/smrt-users@0.27.42
+
 ## 0.27.41
 
 ### Patch Changes