+++ packages/config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @happyvertical/smrt-config
 
+## 1.0.0
+
+### Patch Changes
+
+- @happyvertical/smrt-types@1.0.0
+
+## 0.27.42
+
+### Patch Changes
+
+- @happyvertical/smrt-types@0.27.42
+
 ## 0.27.41
 
 ### Patch Changes