+++ packages/features/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @happyvertical/smrt-features
 
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@1.0.0
+  - @happyvertical/smrt-users@1.0.0
+
+## 0.27.42
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.27.42
+  - @happyvertical/smrt-users@0.27.42
+
 ## 0.27.41
 
 ### Patch Changes