+++ packages/smrt-playground/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @happyvertical/smrt-playground
 
+## 1.0.0
+
+### Patch Changes
+
+- @happyvertical/smrt-svelte@1.0.0
+
+## 0.27.42
+
+### Patch Changes
+
+- @happyvertical/smrt-svelte@0.27.42
+
 ## 0.27.41
 
 ### Patch Changes