+++ packages/social/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @happyvertical/smrt-social
 
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@1.0.0
+  - @happyvertical/smrt-content@1.0.0
+  - @happyvertical/smrt-secrets@1.0.0
+  - @happyvertical/smrt-tenancy@1.0.0
+  - @happyvertical/smrt-video@1.0.0
+  - @happyvertical/smrt-config@1.0.0
+
+## 0.27.42
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.27.42
+  - @happyvertical/smrt-content@0.27.42
+  - @happyvertical/smrt-secrets@0.27.42
+  - @happyvertical/smrt-tenancy@0.27.42
+  - @happyvertical/smrt-video@0.27.42
+  - @happyvertical/smrt-config@0.27.42
+
 ## 0.27.41
 
 ### Patch Changes